+++ src/components/Group.tsx
@@ -1,0 +1,10 @@
+
+export function FormtoolsGroup({ title, children }: {
+	title?: string,
+	children: React.ReactNode
+}) {
+	return <fieldset>
+		{ title && <legend>{title}</legend> }
+		{children}
+	</fieldset>
+}
