+++ src/components/index.ts
@@ -8,3 +8,5 @@ export * from './Checkbox'
 export * from './Radio'
 export * from './File'
 export * from './TagList'
+export * from './Group'
+export * from './Schema'
