+++ src/types/schema.ts
@@ -1,0 +1,7 @@
+import { DefaultProps } from './input'
+
+export interface SchemaProps {
+	data: Array<
+		DefaultProps & { [x: string]: any }
+	>
+}
